+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.11.1] - Unreleased
+
+### Fixed
+
+- Fixed basic handling of nested lists in collab mode.
+
 ## [2.11.0] - 2022-02-18
 
 ### Added